+++ TTauri/GUI/Widgets/ButtonWidget.hpp
@@ -38,7 +38,7 @@ public:
         return r;
     }
 
-    void pipelineImagePlaceVertices(gsl::span<GUI::PipelineImage::Vertex>& vertices, size_t& offset) override;
+    void pipelineImagePlaceVertices(gsl::span<GUI::PipelineImage::Vertex>& vertices, int& offset) override;
 
     void handleMouseEvent(GUI::MouseEvent event) override;
